Add viewport zoom with scroll wheel / trackpad pinch support.

+++ brileta/input_events.py
@@ -269,6 +269,6 @@ class MouseButtonUp(MouseState):
 class MouseWheel(InputEvent):
     """The mouse scroll wheel moved."""
 
-    x: int = 0
-    y: int = 0
+    x: float = 0.0
+    y: float = 0.0
     flipped: bool = False
